Clean up SCA capture scripts.
Move capture_aes_random_batch function to capture.py.
Rename simple_capture_traces.py to capture.py.
Delete simple_capture_traces_batch.py.
Update getting_started guide for SCA.
